requirements.txt
accelerate==1.12.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
apache-airflow==2.9.3
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.14.2
apache-airflow-providers-fab==1.2.2
apache-airflow-providers-ftp==3.10.0
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.1
apispec==6.6.1
argcomplete==3.4.0
asgiref==3.8.1
attrs==23.2.0
babel==2.17.0
beautifulsoup4==4.14.3
blinker==1.8.2
cachelib==0.9.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.4.4
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
courlan==1.3.2
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==41.0.7
dateparser==1.2.2
Deprecated==1.2.14
dill==0.4.0
dnspython==2.6.1
docling==2.66.0
docling-core==2.57.0
docling-ibm-models==3.10.3
docling-parse==4.7.2
docutils==0.16
email_validator==2.2.0
et_xmlfile==2.0.0
faiss-cpu==1.13.2
Faker==39.0.0
filelock==3.20.1
filetype==1.2.0
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2023.12.2
google-re2==1.1.20240702
googleapis-common-protos==1.63.2
greenlet==3.3.0
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
htmldate==1.9.4
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.5.3
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jusText==3.0.2
latex2mathml==3.78.1
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==6.0.2
lxml_html_clean==0.4.3
Mako==1.3.5
Markdown==3.10
markdown-it-py==3.0.0
marko==2.2.1
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.3.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.18
networkx==3.6.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
ordered-set==4.1.0
packaging==24.1
pandas==2.3.3
pathspec==0.12.1
pdfminer.six==20251107
pdfplumber==0.11.8
pendulum==3.0.0
pillow==11.3.0
playwright==1.57.0
pluggy==1.5.0
polyfactory==3.2.0
portalocker==3.2.0
prison==0.2.1
protobuf==4.25.3
psutil==6.0.0
pyclipper==1.4.0
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.18.0
PyJWT==2.8.0
pylatexenc==2.10
PyMuPDF==1.26.7
pypdf==6.5.0
pypdfium2==4.30.0
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-nvd3==0.16.0
pymongo==4.10.1
mongoengine==0.29.1
python-pptx==1.0.2
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.1
qdrant-client==1.16.2
rapidocr==3.4.5
referencing==0.35.1
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.2
rpds-py==0.19.0
rtree==1.4.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
semchunk==2.2.2
sentence-transformers==5.2.0
setproctitle==1.3.3
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.6.0
time-machine==2.14.2
tld==0.13.1
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.57.3
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
triton==3.5.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
tzlocal==5.3.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==2.0.7
Werkzeug==2.2.3
wheel==0.45.1
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
xlsxwriter==3.2.9
yarl==1.9.4
zipp==3.19.2
